Add defaultViewport option to connect + launch
In the process, make Viewport a top level module as it is now used in
Puppeteer and Page.
