+++ lib/js/__tests__/puppeteer_test.js
@@ -17,6 +17,7 @@ var Page$BsPuppeteer = require("../src/Page.js");
 var Unit$BsPuppeteer = require("../src/Unit.js");
 var Target$BsPuppeteer = require("../src/Target.js");
 var Coverage$BsPuppeteer = require("../src/Coverage.js");
+var Viewport$BsPuppeteer = require("../src/Viewport.js");
 var Puppeteer$BsPuppeteer = require("../src/Puppeteer.js");
 var CDPSession$BsPuppeteer = require("../src/CDPSession.js");
 var BrowserFetcher$BsPuppeteer = require("../src/BrowserFetcher.js");
@@ -325,7 +326,7 @@ describe("Page", (function () {
                             }));
               }));
         Jest.testPromise(undefined, "emulate()", (function () {
-                var viewport = Page$BsPuppeteer.Viewport[/* make */0](320, 480, 2, undefined, true, undefined, /* () */0);
+                var viewport = Viewport$BsPuppeteer.make(320, 480, 2, undefined, true, undefined, /* () */0);
                 return page[0].emulate({
                               viewport: viewport,
                               userAgent: ""
