+++ lib/js/src/Page.js
@@ -3,28 +3,6 @@
 var Js_primitive = require("bs-platform/lib/js/js_primitive.js");
 var FrameBase$BsPuppeteer = require("./FrameBase.js");
 
-function make(prim, prim$1, prim$2, prim$3, prim$4, prim$5, _) {
-  var tmp = {
-    width: prim,
-    height: prim$1
-  };
-  if (prim$2 !== undefined) {
-    tmp.deviceScaleFactor = Js_primitive.valFromOption(prim$2);
-  }
-  if (prim$3 !== undefined) {
-    tmp.isMobile = Js_primitive.valFromOption(prim$3);
-  }
-  if (prim$4 !== undefined) {
-    tmp.hasTouch = Js_primitive.valFromOption(prim$4);
-  }
-  if (prim$5 !== undefined) {
-    tmp.isLandscape = Js_primitive.valFromOption(prim$5);
-  }
-  return tmp;
-}
-
-var Viewport = /* module */[/* make */make];
-
 function setBypassCSP(enabled, page) {
   return page.setBypassCSP(enabled);
 }
@@ -63,7 +41,6 @@ function waitForResponseUrl(prim, prim$1, prim$2, _) {
   return prim.waitForResponse(prim$1, prim$2 !== undefined ? Js_primitive.valFromOption(prim$2) : undefined);
 }
 
-exports.Viewport = Viewport;
 exports.setBypassCSP = setBypassCSP;
 exports.waitForNavigation = waitForNavigation;
 exports.WaitForRequest = WaitForRequest;
