+++ src/Page.re
@@ -36,24 +36,6 @@ type cookie = {
   "sameSite": Js.undefined(string),
 };
 
-module Viewport = {
-  [@bs.deriving abstract]
-  type t = {
-    width: int,
-    height: int,
-    [@bs.optional]
-    deviceScaleFactor: float,
-    [@bs.optional]
-    isMobile: bool,
-    [@bs.optional]
-    hasTouch: bool,
-    [@bs.optional]
-    isLandscape: bool,
-  };
-
-  let make = t;
-};
-
 type viewport = Viewport.t;
 
 type emulateOptions = {
